+++ webdriver/tests/support/merge_dictionaries.py
