+++ webdriver/tests/window_maximizing.py
@@ -1,5 +1,5 @@
-from support.inline import inline
-from support.asserts import assert_error, assert_success
+from tests.support.inline import inline
+from tests.support.asserts import assert_error, assert_success
 
 alert_doc = inline("<script>window.alert()</script>")
